Add files via upload
This library shows how to use Apache Kafka and FreePascal. This is translation of librdkafka C library. It contain examples for Win64 and Win32. There are 2 ways how to write FreePascal Kafka application. One is C procedural style and the other one is Object Pascal Style.

Future of KafkaGate application is to become gateway between 0mq and Kafka library. Since 0mq is not the broker, idea is to use Kafka as 0mq broker if neccessery.

KafkaPas is simple GUI application which use KafkaClass.pas and fetch or produce messages to some topic